moved some classes and abstracted DeleteDirectoriesPreprocessor into DeleteFilesPreprocessor
