+++ plugin-utilities-plugin/src/main/resources/META-INF/plugin.xml
@@ -11,7 +11,7 @@
     <extensions defaultExtensionNs="com.intellij">
         <!--Add starters here-->
         <appStarter implementation="org.jetbrains.research.pluginUtilities.CliRunnerExample"/>
-        <appStarter implementation="org.jetbrains.research.pluginUtilities.preprocessing.PreprocessJavaStarter"/>
+        <appStarter implementation="org.jetbrains.research.pluginUtilities.preprocessing.PreprocessKotlinJavaStarter"/>
     </extensions>
 
 </idea-plugin>
